refactor(analysis): extract API surface report owner
